tools/from_gcp_video_upload_to_supase/requirements.txt
httpx>=0.25.0
google-api-python-client>=2.0.0
google-auth>=2.0.0
